+++ lib/src/models/models.dart
@@ -1,2 +1,3 @@
+export 'foreground_notification_config.dart';
 export 'location_settings.dart';
 export 'position.dart';